fix: restore screen sharing preview window functionality
• Fixed preview window creation to use real desktopCapturer source IDs instead of synthetic browser detection IDs
• Moved preview window creation from browser detection flow to StreamSelector flow for proper source handling
• Cleaned up obsolete callPopOut code from inAppUI module
• Consolidated all stream logic into screenSharing module by moving StreamSelector from app/streamSelector/
• Added comprehensive screen sharing IPC handlers for preview window management
• Fixed configuration access in preview window creation function
• Removed duplicate and dead code from multiple refactoring iterations

The preview window now opens correctly when screen sharing is initiated through the proper source selection flow, displaying the actual shared content without "invalid state" errors.
