--- app/inAppUI/callPopOut.html
@@ -1,127 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Teams for Linux - Screen Share</title>
-    <style>
-        body {
-            font-family: sans-serif;
-            margin: 0;
-            padding: 0;
-            background-color: #000;
-            color: #fff;
-            height: 100vh;
-            display: flex;
-            flex-direction: column;
-            overflow: hidden;
-            box-sizing: border-box;
-        }
-        
-        
-        #screen-share-video {
-            width: 100%;
-            height: 100%;
-            object-fit: contain;
-            background-color: #000;
-        }
-    </style>
-</head>
-<body>
-    <video id="screen-share-video" autoplay muted></video>
-
-    <script>
-        window.addEventListener('DOMContentLoaded', async () => {
-            const videoElement = document.getElementById('screen-share-video');
-
-            const updateStatusAndStream = async () => {
-                const isActive = await window.electronAPI.getScreenSharingStatus();
-                document.title = isActive ? 'Teams for Linux - Screen is being shared' : 'Teams for Linux - Nothing is being shared';
-
-                if (isActive) {
-                    try {
-                        const sourceId = await window.electronAPI.getScreenShareStream();
-                        const screenSize = await window.electronAPI.getScreenShareScreen();
-                        console.debug('Screen sharing status:', isActive, 'Source ID:', sourceId, 'Screen size:', screenSize);
-                        if (sourceId) {
-                            const stream = await navigator.mediaDevices.getUserMedia({
-                                audio: false,
-                                video: {
-                                    mandatory: {
-                                        chromeMediaSource: 'desktop',
-                                        chromeMediaSourceId: sourceId,
-                                        minWidth: screenSize.width,
-                                        maxWidth: screenSize.width,
-                                        minHeight: screenSize.height,
-                                        maxHeight: screenSize.height
-                                    }
-                                }
-                            });
-                            videoElement.srcObject = stream;
-                            videoElement.style.display = 'block';
-                            
-                            // Wait for video metadata to load to get actual dimensions
-                            videoElement.addEventListener('loadedmetadata', () => {
-                                const videoWidth = videoElement.videoWidth;
-                                const videoHeight = videoElement.videoHeight;
-                                
-                                if (videoWidth && videoHeight) {
-                                    // Calculate thumbnail size while maintaining aspect ratio
-                                    const maxWidth = 400;
-                                    const maxHeight = 300;
-                                    const aspectRatio = videoWidth / videoHeight;
-                                    
-                                    let windowWidth, windowHeight;
-                                    
-                                    // Scale to thumbnail size maintaining aspect ratio
-                                    if (aspectRatio > maxWidth / maxHeight) {
-                                        // Video is wider - fit to max width
-                                        windowWidth = maxWidth;
-                                        windowHeight = maxWidth / aspectRatio;
-                                    } else {
-                                        // Video is taller - fit to max height
-                                        windowHeight = maxHeight;
-                                        windowWidth = maxHeight * aspectRatio;
-                                    }
-                                    
-                                    // Ensure minimum size
-                                    windowWidth = Math.max(windowWidth, 200);
-                                    windowHeight = Math.max(windowHeight, 120);
-                                    
-                                    // Resize window to thumbnail size
-                                    window.electronAPI.resizeWindow({
-                                        width: Math.round(windowWidth),
-                                        height: Math.round(windowHeight)
-                                    });
-                                    
-                                    console.debug('Resizing thumbnail to:', Math.round(windowWidth), 'x', Math.round(windowHeight), 'for video:', videoWidth, 'x', videoHeight);
-                                }
-                            }, { once: true });
-                        } else {
-                            videoElement.srcObject = null;
-                            videoElement.style.display = 'none';
-                        }
-                    } catch (error) {
-                        console.error('Error getting screen share stream:', error);
-                        videoElement.srcObject = null;
-                        videoElement.style.display = 'none';
-                    }
-                } else {
-                    videoElement.srcObject = null;
-                    videoElement.style.display = 'none';
-                }
-            };
-            // Initial status and stream update
-            updateStatusAndStream();
-            
-            // Listen for screen sharing status changes via IPC and update accordingly
-            if (window?.electronAPI && typeof window.electronAPI.onScreenSharingStatusChanged !== 'function') {
-                // Provide a no-op fallback if not defined
-                window.electronAPI.onScreenSharingStatusChanged = function () {};
-            }
-            window.electronAPI?.onScreenSharingStatusChanged(() => {
-                updateStatusAndStream();
-            });
-        });
-    </script>
-</body>
-</html>
