--- app/inAppUI/callPopOutPreload.js
@@ -1,9 +1,0 @@
-const { contextBridge, ipcRenderer } = require('electron');
-
-contextBridge.exposeInMainWorld('electronAPI', {
-    getScreenSharingStatus: () => ipcRenderer.invoke('get-screen-sharing-status'),
-    getScreenShareStream: () => ipcRenderer.invoke('get-screen-share-stream'),
-    getScreenShareScreen: () => ipcRenderer.invoke('get-screen-share-screen'),
-    startScreenShareDisplay: (sourceId) => ipcRenderer.invoke('start-screen-share-display', sourceId),
-    resizeWindow: (dimensions) => ipcRenderer.send('resize-call-pop-out-window', dimensions),
-});
