+++ app/inAppUI/index.js
@@ -2,7 +2,6 @@ const { BrowserWindow, ipcMain } = require("electron");
 const path = require("path");
 
 let inAppUIWindow = null;
-let callPopOutWindowInstance = null;
 
 function createInAppWindow(options) {
   const {
@@ -41,8 +40,6 @@ function createInAppWindow(options) {
     // Handle window closure, e.g., set inAppUIWindow to null if it was the in-app UI
     if (htmlFile === "inAppUI.html") {
       inAppUIWindow = null;
-    } else if (htmlFile === "callPopOut.html") {
-      callPopOutWindowInstance = null;
     }
   });
 
@@ -76,77 +73,10 @@ ipcMain.on("close-in-app-ui-window", () => {
   }
 });
 
-ipcMain.on("close-call-pop-out-window", () => {
-  if (callPopOutWindowInstance && !callPopOutWindowInstance.isDestroyed()) {
-    callPopOutWindowInstance.close();
-  }
-});
 
-ipcMain.on("resize-call-pop-out-window", (event, { width, height }) => {
-  if (callPopOutWindowInstance && !callPopOutWindowInstance.isDestroyed()) {
-    const [currentWidth, currentHeight] = callPopOutWindowInstance.getSize();
-    const [minWidth, minHeight] = callPopOutWindowInstance.getMinimumSize();
 
-    // Ensure we respect minimum sizes and keep window small (thumbnail size)
-    const newWidth = Math.max(minWidth, Math.min(width, 480));
-    const newHeight = Math.max(minHeight, Math.min(height, 360));
 
-    if (newWidth !== currentWidth || newHeight !== currentHeight) {
-      callPopOutWindowInstance.setSize(newWidth, newHeight);
-      callPopOutWindowInstance.center();
-    }
-  }
-});
-
-function createCallPopOutWindow(config) {
-  if (callPopOutWindowInstance && !callPopOutWindowInstance.isDestroyed()) {
-    callPopOutWindowInstance.focus();
-    return;
-  }
-
-  callPopOutWindowInstance = createInAppWindow({
-    width: 320,
-    height: 180,
-    minWidth: 200,
-    minHeight: 120,
-    show: false,
-    resizable: true,
-    alwaysOnTop: config.alwaysOnTop || false,
-    preload: "callPopOutPreload.js",
-    htmlFile: "callPopOut.html",
-    partition: "persist:teams-for-linux-session",
-  });
-
-  callPopOutWindowInstance.on("closed", () => {
-    // Send IPC event to indicate screen sharing has ended
-    try {
-      const { ipcMain } = require("electron");
-      ipcMain.emit("popup-window-closed");
-    } catch (error) {
-      console.error("Error sending popup-window-closed event:", error);
-    }
-    callPopOutWindowInstance = null;
-  });
-
-  callPopOutWindowInstance.once("ready-to-show", () => {
-    try {
-      const { ipcMain } = require("electron");
-      ipcMain.emit("popup-window-opened");
-    } catch (error) {
-      console.error("Error sending popup-window-opened event:", error);
-    }
-  });
-}
-
-function closeCallPopOutWindow() {
-  if (callPopOutWindowInstance && !callPopOutWindowInstance.isDestroyed()) {
-    callPopOutWindowInstance.close();
-    callPopOutWindowInstance = null;
-  }
-}
 
 module.exports = {
   createInAppUIWindow,
-  createCallPopOutWindow,
-  closeCallPopOutWindow,
 };
