+++ app/inAppUI/preload.js
@@ -5,6 +5,5 @@ contextBridge.exposeInMainWorld(
     {
         getAppVersion: () => ipcRenderer.invoke('get-app-version'),
         closeWindow: () => ipcRenderer.send('close-in-app-ui-window'),
-        createCallPopOutWindow: () => ipcRenderer.invoke('create-call-pop-out-window'),
     }
 );
